Link directly without using -L on Mac OS X.
Ignore-this: 914cb73d128377417483796329d9246

This is a workaround for systems using macports. The library directory of
macports contains a version of libiconv that is incompatible with the
system version of libiconv. The binary GHC package expects the system version.
So adding the macports library directory to the linker search path would be a
showstopper.

darcs-hash:20100428040735-13aa7-dda6c481dcaa41f90cf2eb0340ead636c1026562